requirements.txt
pandas==2.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
plotly==5.19.0
openai==1.12.0
dbt-postgres==1.7.4